requirements.txt
openai>=1.35.0
pydantic>=2.7.0
PySide6>=6.6.3,<6.7
python-dotenv>=1.0.1
tenacity>=8.2.3
networkx>=3.1,<3.3
chromadb>=0.5.5
